+++ lib/overcommit/hook/pre_commit/brakeman.rb
@@ -2,9 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `brakeman` against any modified Ruby/Rails files.
   class Brakeman < Base
     def run
-      result = execute(command +
-                       %w[--exit-on-warn --quiet --summary --only-files] +
-                       [applicable_files.join(',')])
+      result = execute(command + [applicable_files.join(',')])
       return :pass if result.success?
 
       [:fail, result.stdout]