+++ lib/overcommit/hook/pre_commit/bundle_check.rb
@@ -9,7 +9,7 @@ module Overcommit::Hook::PreCommit
       ignored_files = execute(%w[git ls-files -o -i --exclude-standard]).stdout.split("\n")
       return :pass if ignored_files.include?(LOCK_FILE)
 
-      result = execute(command + ['check'])
+      result = execute(command)
       unless result.success?
         return :fail, result.stdout
       end