+++ lib/overcommit/hook/pre_commit/chamber_security.rb
@@ -2,7 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `chamber secure` against any modified Chamber settings files
   class ChamberSecurity < Base
     def run
-      result = execute(command + %w[secure --files] + applicable_files)
+      result = execute(command + applicable_files)
 
       return :pass if result.stdout.empty?
       [:fail, "These settings appear to need to be secured but were not: #{result.stdout}"]