+++ lib/overcommit/hook/pre_commit/coffee_lint.rb
@@ -2,7 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `coffeelint` against any modified CoffeeScript files.
   class CoffeeLint < Base
     def run
-      result = execute(command + %w[--quiet] + applicable_files)
+      result = execute(command + applicable_files)
       return :pass if result.success?
 
       [:fail, result.stdout]