+++ lib/overcommit/hook/pre_commit/merge_conflicts.rb
@@ -2,7 +2,7 @@ module Overcommit::Hook::PreCommit
   # Checks for unresolved merge conflicts
   class MergeConflicts < Base
     def run
-      result = execute(%w[grep -IHn ^<<<<<<<\s] + applicable_files)
+      result = execute(command + applicable_files)
 
       unless result.stdout.empty?
         return :fail, "Merge conflict markers detected:\n#{result.stdout}"