+++ lib/overcommit/hook/pre_commit/reek.rb
@@ -2,7 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `reek` against any modified Ruby files.
   class Reek < Base
     def run
-      result = execute(command + %w[--single-line --no-color] + applicable_files)
+      result = execute(command + applicable_files)
       return :pass if result.success?
 
       output = scrub_output(result.stdout + result.stderr)