+++ lib/overcommit/hook/pre_commit/shell_check.rb
@@ -6,7 +6,7 @@ module Overcommit::Hook::PreCommit
     end
 
     def run
-      result = execute(command + %w[--format=gcc] + applicable_files)
+      result = execute(command + applicable_files)
       return :pass if result.success?
 
       extract_messages(