+++ lib/overcommit/hook/pre_commit/trailing_whitespace.rb
@@ -2,7 +2,7 @@ module Overcommit::Hook::PreCommit
   # Checks for trailing whitespace in files.
   class TrailingWhitespace < Base
     def run
-      result = execute(%w[grep -IHn \s$] + applicable_files)
+      result = execute(command + applicable_files)
       unless result.stdout.empty?
         return :fail, "Trailing whitespace detected:\n#{result.stdout}"
       end