+++ lib/overcommit/hook/pre_commit/travis_lint.rb
@@ -2,7 +2,7 @@ module Overcommit::Hook::PreCommit
   # Runs `travis-lint` against any modified Travis CI files.
   class TravisLint < Base
     def run
-      result = execute(command + %w[lint] + applicable_files)
+      result = execute(command + applicable_files)
       return :pass if result.success?
 
       [:fail, (result.stdout + result.stderr).strip]